Update file structure and home layout
