+++ client/image-ml/src/components/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import {Switch, Route} from 'react-router-dom';
-import './App.css';
+import '../App.css';
 import Home from './Home';
 import Function from './Function';
