+++ client/image-ml/src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import {BrowserRouter} from 'react-router-dom';
-import App from './App';
+import App from './components/App';
 
 ReactDOM.render(
   <React.StrictMode>
